+++ src/resources/loadFont/fonts.ts
@@ -1,0 +1,21 @@
+export function allFonts(){
+    return({
+            'Poppins': require('../../../assets/fonts/Poppins-Regular.ttf'),
+            'Poppins-Bold': require('../../../assets/fonts/Poppins-Bold.ttf'),
+            'Poppins-ExtraBold': require('../../../assets/fonts/Poppins-ExtraBold.ttf'),
+            'Poppins-ExtraLight': require('../../../assets/fonts/Poppins-ExtraLight.ttf'),
+            'Poppins-Light': require('../../../assets/fonts/Poppins-Light.ttf'),
+            'Poppins-Medium': require('../../../assets/fonts/Poppins-Medium.ttf'),
+            'Poppins-SemiBold': require('../../../assets/fonts/Poppins-SemiBold.ttf'),
+            'Poppins-Thin': require('../../../assets/fonts/Poppins-Thin.ttf'),
+            'Inter-Black': require('../../../assets/fonts/Inter-Black.ttf'),
+            'Inter-Bold': require('../../../assets/fonts/Inter-Bold.ttf'),
+            'Inter-ExtraBold': require('../../../assets/fonts/Inter-ExtraBold.ttf'),
+            'Inter-ExtraLight': require('../../../assets/fonts/Inter-ExtraLight.ttf'),
+            'Inter-Light': require('../../../assets/fonts/Inter-Light.ttf'),
+            'Inter-Medium': require('../../../assets/fonts/Inter-Medium.ttf'),
+            'Inter-Regular': require('../../../assets/fonts/Inter-Regular.ttf'),
+            'Inter-SemiBold': require('../../../assets/fonts/Inter-SemiBold.ttf'),
+    }
+    )
+}
